--- config.yml
@@ -1,29 +1,0 @@
-url:
-  login: "http://10.255.0.19/drcom/login"
-  logout: "http://10.255.0.19/drcom/logout"  # 不知道为啥要写登出
-
-users:
-  # 单个账户设置
-  - user:
-      userid: 201930****           # 学/工号
-      password: *******            # 密码
-      ISP: 电信                     # 运营商，可选：教职工、电信、联通、移动
-      0MKKey: 123456               # 不知道有啥用，但是有用
-  # 多账号设置，已解除账号登陆设备数量限制，应用场景较少
-#   - user:
-#       userid:  # 学/工号
-#       password: # 密码
-#       ISP:  # 运营商，可选：教职工、电信、联通、移动
-
-# 通知相关，不建议使用 go-cqhttp，默认 Windows 通知
-msg:
-  go-cqhttp:
-    # url 推送到个人QQ: http://127.0.0.1:port/send_private_msg  群：http://127.0.0.1:port/send_group_msg
-    # token 填写在go-cqhttp文件设置的访问密钥
-    # qq 被通知的QQ
-    # go-cqhttp相关API https://docs.go-cqhttp.org/api
-    url: "http://127.0.0.1:port/send_private_msg"
-    token: *****************************
-    qq: "2517421382"
-  win:
-    title: "自动连接🔗"  # Windows 系统通知标题
