+++ requirements.txt
@@ -1,3 +1,1 @@
 requests==2.25.1
-win10toast-click==0.1.2
-
